+++ lib/very_commons.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'src/page.dart';
 export 'src/page_request.dart';
 export 'src/pageable.dart';
 export 'src/slice.dart';